requirements.txt
experta
groq
python-dotenv